backends/python/django-sync/benchmark/pyproject.toml
[tool.poetry]
name = "benchmark"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
django = "*"
djangorestframework = "*"
psycopg2 = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
